requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.0.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
